+++ jobs/jobs.go
@@ -1,22 +1,18 @@
 package jobs
 
 import (
-	"fmt"
 	"log"
 	"net/http"
-	"os"
 	"time"
 
 	"github.com/go-co-op/gocron"
 
 	"github.com/marcus-crane/gunslinger/db"
 	"github.com/marcus-crane/gunslinger/models"
-	"github.com/marcus-crane/gunslinger/utils"
 )
 
 var (
 	CurrentPlaybackItem models.MediaItem
-	STORAGE_DIR         = utils.GetEnv("STORAGE_DIR", "/tmp")
 )
 
 func SetupInBackground(store db.Store) *gocron.Scheduler {
@@ -24,7 +20,7 @@ func SetupInBackground(store db.Store) *gocron.Scheduler {
 
 	client := http.Client{}
 
-	s.Every(1).Seconds().Do(GetCurrentlyPlayingPlex, store, client)
+	// s.Every(1).Seconds().Do(plex.GetCurrentlyPlaying, store, client)
 	s.Every(15).Seconds().Do(GetRecentlyReadManga, store, client) // Rate limit: 90 req/sec
 	s.Every(15).Seconds().Do(GetCurrentlyPlayingSteam, store, client)
 	s.Every(15).Seconds().Do(GetCurrentlyPlayingTrakt, store, client)
@@ -52,17 +48,4 @@ func SetupInBackground(store db.Store) *gocron.Scheduler {
 	log.Print("Jobs scheduled. Scheduler not running yet.")
 
 	return s
-}
-
-func LoadCover(guid string, extension string) (string, error) {
-	img, err := os.ReadFile(fmt.Sprintf("%s/cover.%s.%s", STORAGE_DIR, guid, extension))
-	if err != nil {
-		return "", err
-	}
-	return string(img), nil
-}
-
-func saveCover(guid string, image []byte, extension string) error {
-	os.WriteFile(fmt.Sprintf("%s/current.jpeg", STORAGE_DIR), image, 0644)
-	return os.WriteFile(fmt.Sprintf("%s/cover.%s.%s", STORAGE_DIR, guid, extension), image, 0644)
 }
