--- plex/testdata/status_music.json
@@ -1,105 +1,0 @@
-{
-  "MediaContainer": {
-    "size": 1,
-    "Metadata": [
-      {
-        "addedAt": 1697708404,
-        "duration": 110837,
-        "grandparentGuid": "plex://artist/5d07bfe9403c6402907b039e",
-        "grandparentKey": "/library/metadata/38305",
-        "grandparentRatingKey": "38305",
-        "grandparentThumb": "/library/metadata/38305/thumb/1699768292",
-        "grandparentTitle": "天気予報",
-        "guid": "plex://track/5d0807a7403c64029039f38f",
-        "index": 9,
-        "key": "/library/metadata/38315",
-        "lastViewedAt": 1697746519,
-        "librarySectionID": "4",
-        "librarySectionKey": "/library/sections/4",
-        "librarySectionTitle": "Music",
-        "musicAnalysisVersion": "1",
-        "parentGuid": "plex://album/627f9da0ad4aee8d8b529682",
-        "parentIndex": 1,
-        "parentKey": "/library/metadata/38306",
-        "parentRatingKey": "38306",
-        "parentStudio": "Geometric Lullaby",
-        "parentThumb": "/library/metadata/38306/thumb/1697708433",
-        "parentTitle": "データベースコンパイラ (Database Compiler)",
-        "parentYear": 2021,
-        "ratingCount": 251,
-        "ratingKey": "38315",
-        "sessionKey": "124",
-        "thumb": "/library/metadata/38306/thumb/1697708433",
-        "title": "あなたの夢を運転する",
-        "type": "track",
-        "updatedAt": 1698386921,
-        "viewCount": 1,
-        "viewOffset": 15704,
-        "Media": [
-          {
-            "audioChannels": 2,
-            "audioCodec": "mp3",
-            "bitrate": 320,
-            "container": "mp3",
-            "duration": 110837,
-            "id": "57430",
-            "Part": [
-              {
-                "container": "mp3",
-                "duration": 110837,
-                "file": "/media/music/天気予報/データベースコンパイラ (Database Compiler)/09 あなたの夢を運転する.mp3",
-                "hasThumbnail": "1",
-                "id": "57992",
-                "key": "/library/parts/57992/1697708422/file.mp3",
-                "size": 6487086,
-                "Stream": [
-                  {
-                    "albumGain": "-14.24",
-                    "albumPeak": "1.000000",
-                    "albumRange": "10.660397",
-                    "audioChannelLayout": "stereo",
-                    "bitrate": 320,
-                    "channels": 2,
-                    "codec": "mp3",
-                    "displayTitle": "MP3 (Stereo)",
-                    "extendedDisplayTitle": "MP3 (Stereo)",
-                    "gain": "-14.24",
-                    "id": "133802",
-                    "index": 0,
-                    "loudness": "-14.24",
-                    "lra": "8.05",
-                    "peak": "0.999969",
-                    "samplingRate": 44100,
-                    "selected": true,
-                    "streamType": 2
-                  }
-                ]
-              }
-            ]
-          }
-        ],
-        "User": {
-          "id": "1",
-          "thumb": "https://plex.tv/users/bd082d86df2dd317/avatar?c=1700022594",
-          "title": "utf9k"
-        },
-        "Player": {
-          "address": "172.17.0.1",
-          "device": "macOS",
-          "machineIdentifier": "85072918-52c8-4189-b457-e0bbc155c7b1",
-          "platform": "macOS",
-          "platformVersion": "23.0.0",
-          "product": "Plexamp",
-          "remotePublicAddress": "172.17.0.1",
-          "state": "playing",
-          "title": "skeith",
-          "version": "4.8.4",
-          "local": false,
-          "relayed": false,
-          "secure": true,
-          "userID": 1
-        }
-      }
-    ]
-  }
-}
